+++ app/src/components/Search/SearchForm.tsx
@@ -4,7 +4,7 @@ import { TextField } from "@mui/material";
 
 import { useSearchProvider } from "../../provider/SearchProvider";
 
-export const Search = () => {
+export const SearchForm = () => {
   const { actions } = useSearchProvider();
   const [params] = useSearchParams();